refactor mesh building
